chore: Rename MtCreditsConsumer.kt to MtCreditsService  & add some docs
