+++ backend/api/src/main/kotlin/io/tolgee/api/v2/controllers/machineTranslation/MtCreditsController.kt
@@ -8,7 +8,7 @@ import io.tolgee.security.ProjectHolder
 import io.tolgee.security.authentication.AllowApiAccess
 import io.tolgee.security.authorization.RequiresOrganizationRole
 import io.tolgee.security.authorization.UseDefaultPermissions
-import io.tolgee.service.machineTranslation.mtCreditsConsumption.MtCreditsConsumer
+import io.tolgee.service.machineTranslation.mtCreditsConsumption.MtCreditsService
 import io.tolgee.service.organization.OrganizationService
 import org.springframework.web.bind.annotation.CrossOrigin
 import org.springframework.web.bind.annotation.GetMapping
@@ -22,7 +22,7 @@ import org.springframework.web.bind.annotation.RestController
 @Tag(name = "Machine translation credits")
 class MtCreditsController(
   private val projectHolder: ProjectHolder,
-  private val mtCreditsConsumer: MtCreditsConsumer,
+  private val mtCreditsService: MtCreditsService,
   private val organizationService: OrganizationService,
 ) {
   @GetMapping("/projects/{projectId:\\d+}/machine-translation-credit-balance")
@@ -35,7 +35,7 @@ class MtCreditsController(
   fun getProjectCredits(
     @PathVariable projectId: Long,
   ): CreditBalanceModel {
-    return mtCreditsConsumer.getCreditBalances(projectHolder.project.organizationOwnerId).model
+    return mtCreditsService.getCreditBalances(projectHolder.project.organizationOwnerId).model
   }
 
   @GetMapping("/organizations/{organizationId:\\d+}/machine-translation-credit-balance")
@@ -49,7 +49,7 @@ class MtCreditsController(
     @PathVariable organizationId: Long,
   ): CreditBalanceModel {
     val organization = organizationService.get(organizationId)
-    return mtCreditsConsumer.getCreditBalances(organization.id).model
+    return mtCreditsService.getCreditBalances(organization.id).model
   }
 
   private val MtCreditBalanceDto.model
